# cpr-backend/src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger CPR - OpenAPI 3.0
  description: |-
    Đây là API của website CPR
  version: 1.0.1
servers:
...
paths:
  /api/item/all:
    get:
      tags:
        - Item
      summary: Get all Items
      description: Returns a list of items
      operationId: getAllItems
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Invalid ID supplied
        "404":
          description: Item not found
  /api/item:
    post:
      tags:
        - Item
      summary: Add a new Item
      description: create a new Item
      operationId: createItem
      requestBody:
        description: Create a new item
        content:
          application/json:
            schema:
              required:
                - itemName
                - category
                - itemUrl
              type: object
              properties:
                itemName:
                  type: string
                  example: "Khoai Tây Chiên"
                category:
                  type: array
                  items:
                    type: string
                    example: "Món ăn kèm"
                itemUrl:
                  type: string
                  example: ""
                description:
                  type: string
                  example: Khoai tây chiên ngoài giòn trong mềm
                price:
                  type: integer
                  default: 0
        required: true
      responses:
        "201":
          description: Create item success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
    get:
      tags:
        - Item
      summary: Get all Items with searchTerm and paging
      description: Returns a list of items
      operationId: getAllItemsWithPaging
      parameters:
        - name: searchTerm
          in: query
          description: The search term for items
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: The page number for items
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: The limit number for items
          required: false
          schema:
            type: integer
            default: 5
        - name: status
          in: query
          description: The status of items
          required: false
          schema:
            type: string
            enum:
              - Active
              - Inactive
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Invalid ID supplied
        "404":
          description: Item not found
  /api/item/{itemName}:
    put:
      tags:
        - Item
      summary: Update an existing Item
      description: Update an existing Item by name
      operationId: updateItemById
      parameters:
        - name: itemName
          in: path
          description: name of an Item
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent Item
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Item"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Invalid itemName supplied
        "404":
          description: Item not found
        "422":
          description: Validation exception
    delete:
      tags:
        - Item
      summary: Deletes an Item
      description: delete an Item by name
      operationId: deletePet
      parameters:
        - name: itemName
          in: path
          description: name of an Item
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
        "204":
          description: Delete item success
        "400":
          description: Invalid itemName supplies
        "404":
          description: Item not found
  /api/item/{term}:
    get:
      tags:
        - Item
      summary: Find Item by Name or Category
      description: Returns a list of item
      operationId: getItemByTerm
      parameters:
        - name: term
          in: path
          description: The name or category of items
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Invalid ID supplied
        "404":
          description: Item not found
...
  /api/order/all:
    get:
      tags:
        - Order
      summary: Get all Orders
      description: Returns a list of orders
      operationId: getAllOrders
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid Id supplied
        "404":
          description: Order not found
  /api/order:
    post:
      tags:
        - Order
      summary: Add a new Order
      description: create a new Order
      operationId: createOrder
      requestBody:
        description: Create a new order
        content:
          application/json:
            schema:
              required:
                - orderId
                - username
                - Items
              type: object
              properties:
                orderId:
                  type: string
                  example: 1
                username:
                  type: string
                  example: Tiến Tài
                orderDate:
                  type: string
                  example: ""
                items:
                  type: object
                  properties:
                    itemName:
                      type: string
                      default: Gà Rán
                    quantity:
                      type: integer
                      default: 1
                total:
                  type: integer
                  default: 0
        required: true
      responses:
        "201":
          description: Create order success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
    get:
      tags:
        - Order
      summary: Get all Orders with searchTerm and paging
      description: Returns a list of orders
      operationId: getAllOrdersWithPaging
      parameters:
        - name: searchTerm
          in: query
          description: The search term for orders
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: The page number for orders
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: The limit number for orders
          required: false
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid Id supplied
        "404":
          description: Order not found
  /api/order/{orderId}:
    get:
      tags:
        - Order
      summary: Get Order by Id
      description: Returns an order
      operationId: getOrderById
      parameters:
        - name: orderId
          in: path
          description: Id of an Order
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid Id supplied
        "404":
          description: Order not found
    put:
      tags:
        - Order
      summary: Update an existing Order
      description: Update an existing Order by Id
      operationId: updateOrderById
      parameters:
        - name: orderId
          in: path
          description: Id of an Order
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent Order
        content:
          application/json:
            schema:
              required:
                - username
                - Items
              type: object
              properties:
                username:
                  type: string
                  example: Tiến Tài
                orderDate:
                  type: string
                  example: ""
                items:
                  type: object
                  properties:
                    itemName:
                      type: string
                      default: Gà Rán
                    quantity:
                      type: integer
                      default: 1
                total:
                  type: integer
                  default: 0
                status:
                  type: string
                  description: Order Status
                  example: Success
                  enum:
                    - Success
                    - Cancel
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid Id supplied
        "404":
          description: Order not found
        "422":
          description: Validation exception
    delete:
      tags:
        - Order
      summary: Deletes an Order
      description: delete an Order by Id
      operationId: deleteOrder
      parameters:
        - name: orderId
          in: path
          description: Id of an Order
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
        "204":
          description: Delete order success
        "400":
          description: Invalid ID supplies
        "404":
          description: Order not found
...
components:
  schemas:
    Order:
      type: object
      properties:
        orderId:
          type: string
          example: 1
        username:
          type: string
          example: Tiến Tài
        orderDate:
          type: date
          example: 01/01/2024
        Items:
          type: object
          properties:
            itemName:
              type: string
              default: Gà Rán
            quantity:
              type: integer
              default: 1
        total:
          type: integer
        status:
          type: string
          description: Order Status
          example: Success
          enum:
            - Success
            - Cancel
...
    Item:
      required:
        - itemName
        - category
        - itemUrl
      type: object
      properties:
        itemName:
          type: string
          example: "Khoai Tây Chiên"
        category:
          type: array
          items:
            type: string
            example: "Món ăn kèm"
        itemUrl:
          type: string
          default: ""
        description:
          type: string
          example: "Mô tả món ăn"
        price:
          type: integer
        status:
          type: string
          description: Item Status
          example: Active
          enum:
            - Active
            - Inactive
...